Hide 'CustomString' class into the implementation
